requirements.txt
openai>=1.0.0
python-dotenv>=0.19.0
requests>=2.31.0
numpy>=1.24.0
pandas>=2.0.0
pygame>=2.5.0
pillow>=10.0.0